fpga: pr_decoupler: Add AXI firewall support for region master IF
AXI firewall is working in the SI mode, in which
transactions on the SI side (from master IF of PR region)
are monitored and isolated from the MI side (e.g., DDR interface)
in case of an invalid transaction occurs.

Invalid AXI transactions may be sent from
partial reconfiguration regions during either normal execution
or partial configurations.

When the AXI firewall is blocked (invalid transactions occured),
the MI side is triggered to process the pending transactions.
The AXI firewall MUST be unblocked after all pending transactions
are flushed from the MI side.

As a result, before releasing PR decoupler, this driver must wait for
the corresponding AXI firewall to finish all pending transactions
on the MI side.